+++ noodles-cram/src/io/writer/builder.rs
@@ -104,15 +104,6 @@ impl Builder {
         File::create(dst).map(|file| self.build_from_writer(file))
     }
 
-    /// Builds a CRAM writer from a path.
-    #[deprecated(since = "0.68.0", note = "Use `Builder::build_from_path` instead.")]
-    pub fn build_with_path<P>(self, dst: P) -> io::Result<Writer<File>>
-    where
-        P: AsRef<Path>,
-    {
-        self.build_from_path(dst)
-    }
-
     /// Builds a CRAM writer from a writer.
     ///
     /// # Examples
@@ -136,15 +127,6 @@ impl Builder {
             records: Vec::with_capacity(RECORDS_PER_CONTAINER),
             record_counter: 0,
         }
-    }
-
-    /// Builds a CRAM writer from a writer.
-    #[deprecated(since = "0.68.0", note = "Use `Builder::build_from_writer` instead.")]
-    pub fn build_with_writer<W>(self, writer: W) -> Writer<W>
-    where
-        W: Write,
-    {
-        self.build_from_writer(writer)
     }
 }